Pivoting Technique - In Hindi
Pivoting Technique - In Hindi

What is Pivoting?

Pivoting is a technique that uses to route the traffic from a hacked computer toward other networks that are not accessible by a hacker machine.
Let’s take a scenario to understand how Pivoting works. 
Assume we have two networks −

A network with the range 192.168.1.0/24 where the hacker machine has access, and Another network with the range 10.10.10.0/24. It is an internal network and the hacker doesn’t have access to it.

The hacker will try to hack the second network this machine that has access to both networks to exploit and hack other internal machines.
In this scenario, a hacker will first break into the first network and then use it as a staging point to exploit and hack the internal machines of the second network. This process is known as pivoting because the hacker is using the first network as a pivot to get access to the second network.

*****Prerequisite*******
In this, I am using the Kali Linux machine as an attacker machine, in it, you have to install the sshuttle package (apt-get install sshuttle) on your system. Required credential of victim machine which we utilized for creating a pivoting setup. 
*************************

Linkedin link URL:  https://www.linkedin.com/in/saksham-dixit-a025a72a/

Blog URL: https://www.sakshamdixit.com

Youtube Link: https://www.youtube.com/channel/UCuizqimEZRXT-0yh4bJeKNw

Github Link: https://github.com/pentest-diaries

@SAKSHAM DIXIT